% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do-something.R
\name{print_charlie_if}
\alias{print_charlie_if}
\title{print_charlie_if}
\usage{
print_charlie_if(x, y)
}
\description{
print_charlie_if
}
\examples{
print_charlie_if("hyeyoung", "young")
print_charlie_if("konrad", "siek")
}
